+++ pylatexenc/__init__.py
@@ -35,3 +35,5 @@ Main Site:
 from .version import version_str as _version_str
 
 __version__ = _version_str
+
+